Split saved_objects/lib into two folders and separate file per function (#32914)
