+++ src/legacy/server/saved_objects/export/sort_objects.ts
@@ -18,54 +18,7 @@
  */
 
 import Boom from 'boom';
-import { SavedObject, SavedObjectsClient } from '../service/saved_objects_client';
-
-interface ObjectToExport {
-  id: string;
-  type: string;
-}
-
-interface ExportObjectsOptions {
-  types?: string[];
-  objects?: ObjectToExport[];
-  savedObjectsClient: SavedObjectsClient;
-  exportSizeLimit: number;
-}
-
-export async function getSortedObjectsForExport({
-  types,
-  objects,
-  savedObjectsClient,
-  exportSizeLimit,
-}: ExportObjectsOptions) {
-  let objectsToExport: SavedObject[] = [];
-  if (objects) {
-    if (objects.length > exportSizeLimit) {
-      throw Boom.badRequest(`Can't export more than ${exportSizeLimit} objects`);
-    }
-    ({ saved_objects: objectsToExport } = await savedObjectsClient.bulkGet(objects));
-    const erroredObjects = objectsToExport.filter(obj => !!obj.error);
-    if (erroredObjects.length) {
-      const err = Boom.badRequest();
-      err.output.payload.attributes = {
-        objects: erroredObjects,
-      };
-      throw err;
-    }
-  } else {
-    const findResponse = await savedObjectsClient.find({
-      type: types,
-      sortField: '_id',
-      sortOrder: 'asc',
-      perPage: exportSizeLimit,
-    });
-    if (findResponse.total > exportSizeLimit) {
-      throw Boom.badRequest(`Can't export more than ${exportSizeLimit} objects`);
-    }
-    ({ saved_objects: objectsToExport } = findResponse);
-  }
-  return sortObjects(objectsToExport);
-}
+import { SavedObject } from '../service/saved_objects_client';
 
 export function sortObjects(savedObjects: SavedObject[]) {
   const path = new Set();
