+++ src/legacy/server/saved_objects/import/create_objects_filter.ts
@@ -1,0 +1,54 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { SavedObject } from '../service';
+
+export function createObjectsFilter(
+  skips: Array<{
+    type: string;
+    id: string;
+  }>,
+  overwrites: Array<{
+    type: string;
+    id: string;
+  }>,
+  replaceReferences: Array<{
+    type: string;
+    from: string;
+    to: string;
+  }>
+) {
+  const refReplacements = replaceReferences.map(ref => `${ref.type}:${ref.from}`);
+  return (obj: SavedObject) => {
+    if (skips.some(skipObj => skipObj.type === obj.type && skipObj.id === obj.id)) {
+      return false;
+    }
+    if (
+      overwrites.some(overwriteObj => overwriteObj.type === obj.type && overwriteObj.id === obj.id)
+    ) {
+      return true;
+    }
+    for (const reference of obj.references || []) {
+      if (refReplacements.includes(`${reference.type}:${reference.id}`)) {
+        return true;
+      }
+    }
+    return false;
+  };
+}
